s3 list buckets cmd
