+++ apps/express-server/src/types/index.ts
@@ -1,1 +1,2 @@
+export * from './redis';
 export * from './request';
