+++ apps/express-server/src/types/redis.ts
@@ -1,0 +1,14 @@
+export type SScanResult = {
+  cursor: number;
+  members: string[]
+};
+
+export type SortedSetMember = {
+  score: number;
+  value: string;
+};
+
+export type ZScanResult = {
+  cursor: number;
+  members: SortedSetMember[];
+}
