+++ apps/express-server/src/utils/index.ts
@@ -1,4 +1,6 @@
+export * from './aws';
 export * from './logger';
 export * from './mixpanel';
+export * from './redis';
 export * from './request';
 export * from './validation';
